requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.6.0
baron==0.10.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==4.0
django-appconf==1.0.5
django-cache-url==3.2.3
django-cors-headers==3.10.1
django-environs==0.1.1
django-imagekit==4.1.0
djangorestframework==3.12.4
environs==9.3.5
marshmallow==3.14.1
pilkit==2.0
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
python-dotenv==0.19.2
pytz==2021.3
redbaron==0.9.2
rply==0.7.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
